requirements.txt
tweepy==3.10.0
Mastodon.py==1.6.1
python-dotenv==0.18.0
